Add libcxx Compatability Mode
At the moment, AnyBlob can only link against the gnu standard library.
This commit adds a new CMake option `ANYBLOB_LIBCXX_COMPAT` that allows
building AnyBlob and linking against libcxx.

The main changes in this commit are:
- Missing header includes. We don't bring in the same transitive headers
  when including the libcxx headers, as a result some string headers
  were missing in some places.
- Remove `ThroughputResolver` when building with libcxx. This is the
  more invasive of the two changes. The `ThroughputResolver` depends on
  `__gnu_pbds::tree`, which is not spported in libcxx. When building in
  libcxx compatibility mode, we now remove this resolve completely. In
  this case, the `io_uring_socket.cpp` falls back onto the default
  resolver.

We have verified that we can build all targets in libcxx compatibility
mode.

We have verified that both unit and integration tests pass when building
with `clang-15` and the libcxx compatibility mode.