Add Clang Toolchain to the CI
This commit adds the clang toolchain to the CI. We turn the main jobs
into a test matrix that either runs with the clang, or gcc toolchain.

When using the gcc toolchain everything stays as-is. When using the
clang toolchain we use clang/clang++ as C/C++ compilers, respectively.
We also build `AnyBlob` using `-DANYBLOB_LIBCXX_COMPAT=1`, which ensures
that `libcxx` is used as a standard library.

In principle, the test matrix could be even bigger: we could build with
libcxx and gcc, and gnu c++ and clang. But we don't tests this for now.